LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/script-src-self/worklet-layout.https-expected.txt


FAIL Content Security Policy: Expects allowed for worklet-layout to same-https origin and keep-origin redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'get_worklet(type).addModule')"
PASS Content Security Policy: Expects allowed for worklet-layout to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-layout to same-https origin and no-redirect redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'get_worklet(type).addModule')"
PASS Content Security Policy: Expects allowed for worklet-layout to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-layout to cross-https origin and keep-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for worklet-layout to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worklet-layout to cross-https origin and no-redirect redirection from https context.
FAIL Content Security Policy: Expects blocked for worklet-layout to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worklet-layout to cross-https origin and swap-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for worklet-layout to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worklet-layout to same-https origin and swap-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for worklet-layout to same-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
